Extract main impl to libraries
