+++ .gitignore
@@ -253,4 +253,4 @@ $RECYCLE.BIN/
 
 # End of https://www.toptal.com/developers/gitignore/api/rust,macos,windows,linux,visualstudiocode,emacs,vim,intellij+all
 
-input/*
+input/
