+++ Cargo.toml
@@ -1,10 +1,2 @@
-[package]
-authors = ["Nick Hirakawa <nickhirakawa@gmail.com>"]
-edition = "2018"
-name = "advent-of-code-2020"
-version = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-nom = "6.0.1"
-regex = "1.4.2"
+[workspace]
+members = ["runner", "advent-of-code-2020", "common"]
