+++ advent-of-code-2020/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "advent-of-code-2020"
+version = "0.1.0"
+authors = ["Nick Hirakawa <nickhirakawa@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+common = { path = "../common" }
+nom = "6.0.1"
+regex = "1.4.2"
