+++ advent-of-code-2020/src/day_eight.rs
@@ -1,6 +1,6 @@
 use std::collections::HashSet;
 
-use crate::prelude::*;
+use common::prelude::*;
 use nom::{
     branch::alt,
     bytes::complete::tag,
