+++ advent-of-code-2020/src/day_eighteen.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use common::prelude::*;
 use nom::{
     branch::alt,
     bytes::complete::tag,
