+++ advent-of-code-2020/src/day_eleven.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use common::prelude::*;
 use std::collections::HashMap;
 
 type Coordinate = (i32, i32);
