+++ advent-of-code-2020/src/day_nineteen.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::prelude::*;
+use common::prelude::*;
 use nom::{
     branch::alt,
     bytes::complete::tag,
