+++ advent-of-code-2020/src/day_one.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use common::prelude::*;
 
 pub fn run() -> AdventOfCodeResult<u64, u64> {
     let expenses = read_expenses()?;
