+++ advent-of-code-2020/src/day_seven.rs
@@ -1,4 +1,4 @@
-use crate::answer::{AdventOfCodeError, AdventOfCodeResult, PartAnswer};
+use common::prelude::*;
 use nom::{
     branch::alt,
     bytes::complete::tag,
