+++ advent-of-code-2020/src/day_seventeen.rs
@@ -1,7 +1,7 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::iter::IntoIterator;
 
-use crate::prelude::*;
+use common::prelude::*;
 
 pub fn run() -> AdventOfCodeResult<u64, u64> {
     let input = include_str!("../input/day-17.txt");
