+++ advent-of-code-2020/src/day_twenty_two.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 use std::collections::VecDeque;
 
-use crate::prelude::*;
+use common::prelude::*;
 use nom::{
     branch::alt,
     bytes::complete::tag,
