+++ advent-of-code-2020/src/day_two.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use common::prelude::*;
 use nom::{
     bytes::complete::tag,
     character::complete::{alpha1, anychar, digit1, newline},
