+++ common/src/lib.rs
@@ -1,0 +1,2 @@
+mod answer;
+pub mod prelude;
