+++ runner/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+authors = ["Nick Hirakawa <nickhirakawa@gmail.com>"]
+edition = "2018"
+name = "runner"
+version = "0.1.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+advent-of-code-2020 = { path = "../advent-of-code-2020" }
+common = { path = "../common" }
+nom = "6.0.1"
+regex = "1.4.2"
